feat: Add alarm clock functionality
This commit adds the code for an alarm clock functionality to the repository. The code allows users to set an alarm time and plays a sound when the set time is reached. The alarm clock interface is implemented using Tkinter library. The code also uses threading to run the alarm in the background while the main program continues to run.